# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QUEPON GGX Command System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerPlayer
      tags: [auth]
      summary: Register a new player account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login as player or admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # --- PCs ---
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List all feedback (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"

    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback (player)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
  /feedback/{feedbackId}/resolve:
    patch:
      operationId: resolveFeedback
      tags: [feedback]
      summary: Mark feedback as resolved (admin)
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackResolveInput"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  # --- PLAYERS (admin) ---
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List available menu items
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{itemId}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- ORDERS ---
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, player sees own)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a menu order (player)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # --- DASHBOARD ---
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        displayName:
          type: string
        phone:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, role, status, walletBalance, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [player, admin, superAdmin]
        status:
          type: string
          enum: [active, disabled, banned]
        walletBalance:
          type: number
        totalSpent:
          type: number
        totalPlayTimeMinutes:
          type: integer
        sessionCount:
          type: integer
        createdAt:
          type: string

    # --- PCs ---
...
    Feedback:
      type: object
      required: [id, category, message, status, isAnonymous, createdAt]
      properties:
        id:
          type: string
        userId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        category:
          type: string
          enum: [issue, suggestion, general, complaint, featureRequest]
        message:
          type: string
        relatedPcId:
          type: ["string", "null"]
        relatedSessionId:
          type: ["string", "null"]
        isAnonymous:
          type: boolean
        status:
          type: string
          enum: [open, reviewing, resolved, closed, escalated]
        assignedTo:
          type: ["string", "null"]
        resolvedBy:
          type: ["string", "null"]
        resolutionNote:
          type: ["string", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    FeedbackInput:
      type: object
      required: [category, message]
      properties:
        category:
          type: string
        message:
          type: string
        relatedPcId:
          type: ["string", "null"]
        relatedSessionId:
          type: ["string", "null"]
        isAnonymous:
          type: boolean
    FeedbackResolveInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
        resolutionNote:
          type: ["string", "null"]

    # --- PLAYERS ---
...
    MenuItem:
      type: object
      required: [id, name, category, price, isAvailable, displayOrder, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [drinks, snacks, meals, services, others]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
    MenuItemInput:
      type: object
      required: [name, category, price]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
        displayOrder:
          type: integer
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        price:
          type: number
        isAvailable:
          type: boolean
        displayOrder:
          type: integer

    # --- ORDERS ---
    OrderItem:
      type: object
      required: [itemId, name, quantity, price, subtotal]
      properties:
        itemId:
          type: string
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    Order:
      type: object
      required: [id, userId, username, status, items, totalAmount, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        sessionId:
          type: ["string", "null"]
        pcId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, preparing, served, cancelled, rejected]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        paymentSource:
          type: string
        acceptedBy:
          type: ["string", "null"]
        servedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        servedAt:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [items]
      properties:
        sessionId:
          type: ["string", "null"]
        pcId:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            required: [itemId, quantity]
            properties:
              itemId:
                type: string
              quantity:
                type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, preparing, served, cancelled, rejected]

    # --- DASHBOARD ---
...