# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QUEPON GGX Command System API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackCategory } from "./feedbackCategory";
import type { FeedbackStatus } from "./feedbackStatus";

export interface Feedback {
  id: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  username?: string | null;
  category: FeedbackCategory;
  message: string;
  /** @nullable */
  relatedPcId?: string | null;
  /** @nullable */
  relatedSessionId?: string | null;
  isAnonymous: boolean;
  status: FeedbackStatus;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  resolvedBy?: string | null;
  /** @nullable */
  resolutionNote?: string | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}