# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QUEPON GGX Command System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  userId: string;
  username: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  pcId?: string | null;
  status: OrderStatus;
  items: OrderItem[];
  totalAmount: number;
  paymentSource?: string;
  /** @nullable */
  acceptedBy?: string | null;
  /** @nullable */
  servedBy?: string | null;
  createdAt: string;
  /** @nullable */
  acceptedAt?: string | null;
  /** @nullable */
  servedAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
}