# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QUEPON GGX Command System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  status: UserStatus;
  walletBalance: number;
  totalSpent?: number;
  totalPlayTimeMinutes?: number;
  sessionCount?: number;
  createdAt: string;
}